// src/use-data-store/types.ts
export type HandledEvents = 'created' | 'patched' | 'updated' | 'removed'
export type HandleEventsFunction<M extends AnyData> = (item: M, ctx: { model: M; models: any }) => any

export type HandleEvents<M extends AnyData> =
  | {
      [event in HandledEvents]: HandleEventsFunction<M>
    }
  | boolean